compute requirejs config baseUrl on client-side
